safe-fire-ai1/scamshield/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
python-whois>=0.9.0
requests>=2.31.0
pillow>=10.0.0
pytesseract>=0.3.10
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
dnspython>=2.4.0
python-dotenv>=1.0.0
slowapi>=0.1.9
email-validator>=2.0.0
structlog>=24.0.0
sentry-sdk>=1.40.0
twilio>=8.0.0
